test(action-sheet): update basic tests
